+++ src/components/datetime.js
@@ -1,18 +1,12 @@
 import React from 'react';
 import { TextField } from '@mui/material';
+import { getTimezoneOffset } from '../functions';
 
-function getTimezoneOffset(timezone) {
-    const date = new Date();
-    const utcDate = new Date(date.toLocaleString('en-US', { timeZone: 'UTC' }));
-    const tzDate = new Date(date.toLocaleString('en-US', { timeZone: timezone }));
-    return (utcDate - tzDate) / (1000 * 60);
-}
+var vars = require("../variables");
 
 // all in seconds, not milliseconds
 const DateTimeField = ({ label, defaultValue, onChange, fullWidth = false, sx = {}, disabled = false }) => {
-    const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone; // this could be configured in settings soon
-
-    const defaultValueConverted = defaultValue !== undefined ? new Date(new Date(defaultValue * 1000).getTime() - getTimezoneOffset(timezone) * 60000).toISOString().slice(0, 16) : undefined; // ISO gives UTC time, we need to calculate the timezone offset
+    const defaultValueConverted = defaultValue !== undefined ? new Date(new Date(defaultValue * 1000).getTime() - getTimezoneOffset(vars.userSettings.display_timezone) * 60000).toISOString().slice(0, 16) : undefined; // ISO gives UTC time, we need to calculate the timezone offset
 
     const handleChange = (event) => {
         const newTimestamp = new Date(event.target.value).getTime() / 1000;